Fixing support for LED String in Rainbow pattern for IPSTube
Currently only 6 LEDs are counted in the rainbow pattern for IPSTube because of that variable in the loop.
The phase_per_digit is properly calculated based on the setting from platformio.ini but the loop is smaller. Extending that to NUM_BACKLIGHT_LEDS fixes the issue and makes the strip work as expected.